fix: VNPay payment integration & currency display - Added USD to VND conversion (rate 25,000) for car purchase payments - Fixed VNPay amount calculation and redirect logic - Updated currency display from VND to USD in car listings and buy forms - Fixed admin order details to show USD - Added VNPay credentials from .env - Reverted rental car display to VND

+++ app/Http/Controllers/frontend/AccessoryOrderController.php
@@ -126,7 +126,7 @@ class AccessoryOrderController extends Controller
             $inputData = [
                 "vnp_Version" => "2.1.0",
                 "vnp_TmnCode" => $vnp_TmnCode,
-                "vnp_Amount" => (float)$totalPrice * 100, // Amount in VND cents
+                "vnp_Amount" => (int)($totalPrice * 100), // Amount in VND cents
                 "vnp_Command" => "pay",
                 "vnp_CreateDate" => date('YmdHis'),
                 "vnp_CurrCode" => "VND",
